code cleaning
HashMap --> Map/Hashtable
Vector --> Array/ArrayList
Iterator --> for(entry:Map) 
unnecessary casts in (xxx) x.get removed

git-svn-id: https://svn.code.sf.net/p/jspecview/svn@415 1262bc75-0126-4021-b206-820d13540e36

+++ dev2/JSpecView/src/jspecview/applet/JSVApplet.java
@@ -1494,7 +1494,7 @@ public class JSVApplet extends JApplet implements PeakPickedListener {
       if (JSpecViewUtils.DEBUG) {
         System.out.println("KEY-> " + key + " VALUE-> " + value);
       }
-      Integer iparam = (Integer) htParams.get(key);
+      Integer iparam = htParams.get(key);
       try {
         switch (iparam == null ? -1 : iparam.intValue()) {
         case -1:
